* Optimize for size * Remove static healthcheck and just rely on the CommandBox image healthcheck, which makes more sense * Removed some debugging files * Updated CommandBox to 4.0.0 * `ORM_DIALECT` was not being used in the right place. * `EXPRESS` and `HEADLESS` was not evaluated to a boolean, just existence * Removed usage of lowercase `install`, `be`, and `express`, you must use uppercase env variables.

+++ build/run-contentbox.sh
@@ -2,10 +2,11 @@
 set -e
 
 cd $APP_DIR
+
 #######################################################################################
 # Express Edition
 #######################################################################################
-if [[ $express ]] || [[ $EXPRESS ]]; then
+if [[ $EXPRESS ]] && [[ $EXPRESS == true ]]; then
 
 	echo "INFO: Express installation specified.  Configuring H2 Database."
 
@@ -36,7 +37,7 @@ fi
 # INSTALLER
 # If our installer flag has not been passed, then remove that module
 #######################################################################################
-if [[ ! $installer ]] && [[ ! $install ]] && [[ ! $INSTALL ]]; then
+if [[ ! $installer ]] && [[ ! $INSTALL ]]; then
 	echo "INFO: Removing installer..."
 	rm -rf ${APP_DIR}/modules/contentbox-installer
 fi
@@ -45,7 +46,7 @@ fi
 # HEADLESS
 # If true, then remove the contentbox admin
 #######################################################################################
-if [[ $HEADLESS ]]; then
+if [[ $HEADLESS ]] && [[ $HEADLESS == true ]]; then
 	echo "INFO: Removing admin module..."
 	rm -rf ${APP_DIR}/modules/contentbox/modules/contentbox-admin
 fi
